+++ src/Models/EntryCategory.php
@@ -252,6 +252,11 @@ class EntryCategory extends \Hirtz\Skeleton\Db\ActiveRecord implements TrailMode
         return false;
     }
 
+    public function getPermissionName(): string
+    {
+        return Entry::AUTH_ENTRY;
+    }
+
     public function getAdminName(): string
     {
         return Yii::t('cms', 'ENTRY_CATEGORY_TRAIL_NAME');